# src/main/resources/courtregister-openapi.yaml
...
openapi: 3.0.3
info:
  title: Court Register operations API
  version: 1.0.0
  description: >-
    The named operator actions that replaced the operations CLI in increment 005. Every endpoint is
    behind cp-auth-rules-filter, admitting the group "Second Line Support" and no other, with the
    caller's identity taken from the CJSCPPUID header; every endpoint is audited by
    cp-audit-filter-springboot. No response carries defendant detail, an unmasked recipient address,
    a payload, a register document, an exception message, or any value the caller supplied.
servers:
  - url: /
    description: The pod itself, reached only through the internal gateway.
...
paths:
...
  /operations/batches:
    get:
      tags: [batches]
      operationId: listBatches
      summary: The batches of one register date
      parameters:
        - name: date
          in: query
          required: true
          description: The register date, as an ISO local date.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: That date's batches, each with its masked recipients.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchListingResponse"
        "400":
          $ref: "#/components/responses/Refused"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Unauthorised"
        "501":
          $ref: "#/components/responses/NotWired"
        "503":
          $ref: "#/components/responses/Refused"
  /operations/batches/generate:
    post:
      tags: [batches]
      operationId: generateRegister
      summary: Regenerate a register date
      description: >-
        Asynchronous. It validates the request, reads the cutover flag through the same gate the
        18:00 run reads it through, records a run id and answers 202; the work runs on the
        generation scheduler's single thread holding the register-generation lock. What the run did
        is read from the run report line - run_id, trigger=operator, reason=overridden where the
        flag was overridden - and from GET /operations/batches. SCHEDULE_RUNNING and
        generation-failed are recorded as the background run's outcome and are never a status here.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRegisterRequest"
      responses:
        "202":
          description: Accepted; the run id the run line is correlated by.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateRegisterResponse"
        "400":
          $ref: "#/components/responses/Refused"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Unauthorised"
        "409":
          $ref: "#/components/responses/Refused"
        "501":
          $ref: "#/components/responses/NotWired"
        "503":
          $ref: "#/components/responses/NotAudited"
  /operations/batches/{batchId}/notify:
    post:
      tags: [batches]
      operationId: notifyBatch
      summary: Re-request the recipients of one batch no e-mail has been accepted for
      description: >-
        Synchronous, and only the owed recipients. Two concurrent calls are arbitrated by the claim
        RegisterNotifierService already takes; nothing new arbitrates them.
      parameters:
        - name: batchId
          in: path
          required: true
          description: The batch, as a UUID.
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: The batch settled, with the tally.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotifyBatchResponse"
        "400":
          $ref: "#/components/responses/Refused"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Unauthorised"
        "404":
          $ref: "#/components/responses/Refused"
        "409":
          $ref: "#/components/responses/Refused"
        "500":
          $ref: "#/components/responses/Refused"
        "501":
          $ref: "#/components/responses/NotWired"
        "502":
          $ref: "#/components/responses/Refused"
        "503":
          $ref: "#/components/responses/Refused"
        "504":
          $ref: "#/components/responses/Refused"
...
components:
  responses:
    Refused:
      description: >-
        A refusal or a failure, as an RFC 9457 ProblemDetail carrying a bounded reason. Never an
        exception message, never a store's or a far end's own words, never a value the caller sent.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotAudited:
      description: >-
        The call's request event could not be published to the audit context, so it was refused
        rather than made. Being audited is a condition of this surface existing at all
        (constitution Principle III(b)), and the request event is published before the action -
        which is the only moment there is still something to refuse. Carries
        `reason: AUDIT_UNAVAILABLE`. The endpoints that also answer 503 for a store that will not
        answer carry both codes under the one status; the code is what tells them apart.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    NotWired:
      description: >-
        This pod does not hold the generation half, so the action has no beans to reach. Exactly
        what the CLI answered on such a pod; a 404 would read as a mistyped URL and a 500 as a
        bean-definition error reaching an operator.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
    Unauthenticated:
      description: >-
        No CJSCPPUID identity on the request. Produced by cp-auth-rules-filter through sendError, so
        the body is api/OperationsErrorAttributes' - the same shape, without passing the advice.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FilterRefusal"
    Unauthorised:
      description: >-
        The caller is not in "Second Line Support", or the identity service could not be asked, or
        the action has no allow rule. All three are the same answer on purpose: fail-closed is the
        only safe direction for an authorisation decision.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FilterRefusal"
  schemas:
...
    ProblemDetail:
      type: object
      description: >-
        RFC 9457. `detail` is never populated and `instance` is empty on purpose: Spring fills a null
        instance with the request URI, which on the notify path is the batch id the caller typed.
      properties:
        type:
          type: string
        title:
          type: string
          description: The status' own reason phrase, from a closed set.
        status:
          type: integer
          format: int32
        instance:
          type: string
          description: Always empty. The request URI is a value the caller supplied.
        reason:
          $ref: "#/components/schemas/Reason"
        argument:
          type: string
          description: >-
            On a 400 about one argument: this service's own name for it, never the value that was
            sent.
          enum: [date, batchId, recordedBefore, sharedBefore, since]
        batchId:
          type: string
          format: uuid
        runId:
          type: string
        date:
          type: string
          format: date
        released:
          type: integer
          format: int32
          description: The partial tally of a run that stopped part way, where there was one.
        registers:
          type: integer
          format: int32
        batches:
          type: integer
          format: int32
        requested:
          type: integer
          format: int32
        deferred:
          type: integer
          format: int32
        delivered:
          type: object
          additionalProperties:
            type: string
          description: What each sink did, on a report that was built and not delivered.
      required: [status, title, reason]
    FilterRefusal:
      type: object
      description: >-
        What cp-auth-rules-filter's sendError produces, rendered by api/OperationsErrorAttributes:
        the status, a bounded reason and nothing else. No path, no trace, no message - the default
        body echoes the path, which for an unmapped path is a string the caller typed.
      properties:
        status:
          type: integer
          format: int32
        title:
          type: string
        reason:
          type: string
      required: [status]
...
    BatchListingResponse:
      type: object
      properties:
        date:
          type: string
          format: date
        batches:
          type: array
          items:
            $ref: "#/components/schemas/BatchSummary"
      required: [date, batches]
    BatchSummary:
      type: object
      properties:
        batchId:
          type: string
          format: uuid
        courtHouse:
          type: string
          nullable: true
          description: Absent where the batch names no venue; the command printed a dash.
        state:
          $ref: "#/components/schemas/BatchStatus"
        records:
          type: integer
          format: int32
        recipients:
          type: array
          items:
            $ref: "#/components/schemas/RecipientOutcome"
      required: [batchId, state, records, recipients]
...
    GenerateRegisterRequest:
      type: object
      additionalProperties: false
      description: >-
        Closed, for the reason the inbound queue message's contract is closed: an unknown field is
        producer drift, and tolerating it hides the drift until it matters.
      properties:
        date:
          type: string
          format: date
          description: The register date. Required, and never defaulted.
        courtHouse:
          type: string
          description: One court house of that date.
        batchId:
          type: string
          format: uuid
          description: One batch of that date.
        recordedBefore:
          type: string
          format: date-time
          description: Only the registers recorded before this instant.
        ignoreFlag:
          type: boolean
          default: false
          description: >-
            The break-glass, and only with an explicit batchId. Overriding the cutover flag is for
            one batch at a time: re-driving a whole day under override is one call per failed batch,
            deliberately. Recorded in the audit event with the caller's identity and on the run
            report line.
      required: [date]
    GenerateRegisterResponse:
      type: object
      properties:
        runId:
          type: string
        date:
          type: string
          format: date
          description: This service's own parse of the date, not the characters the caller typed.
        overridden:
          type: boolean
      required: [runId, date, overridden]
    NotifyBatchResponse:
      type: object
      properties:
        batchId:
          type: string
          format: uuid
        accepted:
          type: integer
          format: int32
        failed:
          type: integer
          format: int32
        state:
          $ref: "#/components/schemas/BatchStatus"
        disposition:
          type: string
          enum: [settled]
          description: >-
            Only SETTLED is a 200. The other three dispositions are statuses - already-notifying is
            409, claim-lost and incomplete are 500.
      required: [batchId, accepted, failed, state, disposition]
...